requirements.txt
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
charset-normalizer==3.1.0
colorama==0.4.4
coloredlogs==15.0.1
discord-py-slash-command==4.2.1
discord.py==2.2.2
frozenlist==1.3.3
humanfriendly==10.0
idna==3.4
multidict==6.0.4
pycodestyle==2.10.0
python-dotenv==0.21.1
tomli==2.0.1
yarl==1.8.2
